[REFACTOR] 패키지 이전: AvatarTheme -> model
