+++ src/main/java/com/wakeUpTogetUp/togetUp/api/avatar/model/Avatar.java
@@ -1,12 +1,16 @@
 package com.wakeUpTogetUp.togetUp.api.avatar.model;
 
-import com.wakeUpTogetUp.togetUp.api.avatar.AvatarTheme;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.EnumType;
+import javax.persistence.Enumerated;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-
-import javax.persistence.*;
-
 import org.hibernate.annotations.DynamicInsert;
 
 @Setter
